+++ lib/pprint/colours.ts
@@ -31,9 +31,9 @@ export function darkPurple(d: Doc): Doc {
     return colour(Fmt.Colour.DarkPurple, d);
 }
 
-/** Display a document with gold forecolour. */
-export function gold(d: Doc): Doc {
-    return colour(Fmt.Colour.Gold, d);
+/** Display a document with orange forecolour. */
+export function orange(d: Doc): Doc {
+    return colour(Fmt.Colour.Orange, d);
 }
 
 /** Display a document with gray forecolour. */
@@ -56,9 +56,9 @@ export function green(d: Doc): Doc {
     return colour(Fmt.Colour.Green, d);
 }
 
-/** Display a document with aqua forecolour. */
-export function aqua(d: Doc): Doc {
-    return colour(Fmt.Colour.Aqua, d);
+/** Display a document with light blue forecolour. */
+export function lightBlue(d: Doc): Doc {
+    return colour(Fmt.Colour.LightBlue, d);
 }
 
 /** Display a document with red forecolour. */
@@ -66,9 +66,9 @@ export function red(d: Doc): Doc {
     return colour(Fmt.Colour.Red, d);
 }
 
-/** Display a document with light purple forecolour. */
-export function lightPurple(d: Doc): Doc {
-    return colour(Fmt.Colour.LightPurple, d);
+/** Display a document with pink forecolour. */
+export function pink(d: Doc): Doc {
+    return colour(Fmt.Colour.Pink, d);
 }
 
 /** Display a document with yellow forecolour. */
@@ -81,9 +81,44 @@ export function white(d: Doc): Doc {
     return colour(Fmt.Colour.White, d);
 }
 
-/** Display a document with Minecoin gold forecolour. */
-export function minecoinGold(d: Doc): Doc {
-    return colour(Fmt.Colour.MinecoinGold, d);
+/** Display a document with gold forecolour. */
+export function gold(d: Doc): Doc {
+    return colour(Fmt.Colour.Gold, d);
+}
+
+/** Display a document with warm light gray forecolour. */
+export function warmLightGray(d: Doc): Doc {
+    return colour(Fmt.Colour.WarmLightGray, d);
+}
+
+/** Display a document with cool light gray forecolour. */
+export function coolLightGray(d: Doc): Doc {
+    return colour(Fmt.Colour.CoolLightGray, d);
+}
+
+/** Display a document with dark brown forecolour. */
+export function darkBrown(d: Doc): Doc {
+    return colour(Fmt.Colour.DarkBrown, d);
+}
+
+/** Display a document with darker red forecolour. */
+export function darkerRed(d: Doc): Doc {
+    return colour(Fmt.Colour.DarkerRed, d);
+}
+
+/** Display a document with brown forecolour. */
+export function brown(d: Doc): Doc {
+    return colour(Fmt.Colour.Brown, d);
+}
+
+/** Display a document with dark gold forecolour. */
+export function darkGold(d: Doc): Doc {
+    return colour(Fmt.Colour.DarkGold, d);
+}
+
+/** Display a document with aqua forecolour. */
+export function aqua(d: Doc): Doc {
+    return colour(Fmt.Colour.Aqua, d);
 }
 
 /** Display a document with a forecolour given in the first parameter. */